ChatAgent/requirements.txt
# Dependencias para el ChatAgent de PrediRuta
# Python 3.11+

# Framework web
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Google Gemini
google-generativeai==0.3.2

# Validación de datos
pydantic==2.5.3
pydantic-settings==2.1.0

# Variables de entorno
python-dotenv==1.0.0

# Utilidades
httpx==0.26.0
